switch to decimal, code cleanup
the precision of float was making percentage discounts difficult to predict (and a user of the system would probably be unlikely to care).

+++ API/Entities/Product.cs
@@ -1,15 +1,15 @@
 namespace ShoppingCart.Entities;
 
-public class Product(string sku, float unitPrice)
+public class Product(string sku, decimal unitPrice)
 {
     // todo-at: should product SKU be case-sensitive? at the moment it is (case-sensitive)
     public string Sku { get; set; } = sku;
 
-    public float UnitPrice { get; set; } = unitPrice;
+    public decimal UnitPrice { get; set; } = unitPrice;
 
     // todo-at: should a class be extracted here? perhaps
     public string? SpecialPriceName { get; set; } = null;
-    public float? SpecialPrice { get; set; } = null;
+    public decimal? SpecialPrice { get; set; } = null;
 
     public Product Clone() =>
         (Product)this.MemberwiseClone();
